Added choose files; allow to change conda environment

+++ R/install.R
@@ -95,7 +95,7 @@ NULL
 #' @rdname conda-env
 #' @export
 CONDAENV_NAME <- local({
-  name <- "rpymat-conda-env"
+  name <- NULL
   function(env_name){
     if(!missing(env_name)){
       stopifnot(length(env_name) == 1 && is.character(env_name))
@@ -105,6 +105,13 @@ CONDAENV_NAME <- local({
         name <<- sprintf("rpymat-conda-env-%s", env_name)
       }
       message("Environment switched to: ", name)
+    } else if(is.null(name)) {
+      conda_prefix <- trimws(Sys.getenv("R_RPYMAT_CONDA_PREFIX", unset = ""))
+      if( conda_prefix == "" ) {
+        return("rpymat-conda-env")
+      } else {
+        return(sprintf("%s-rpymat-conda-env", basename(conda_prefix)))
+      }
     }
     name
   }
@@ -124,12 +131,24 @@ install_root <- function(){
 #' @rdname conda-env
 #' @export
 conda_path <- function(){
+  conda_exe <- Sys.getenv("R_RPYMAT_CONDA_EXE", unset = "")
+  if( !identical(conda_exe, "") ) {
+    path <- dirname(dirname(conda_exe))
+    if(dir.exists(path)) {
+      return(path)
+    }
+  }
   file.path(install_root(), "miniconda", fsep = "/")
 }
 
 #' @rdname conda-env
 #' @export
 conda_bin <- function(){
+  conda_exe <- Sys.getenv("R_RPYMAT_CONDA_EXE", unset = "")
+  if( !identical(conda_exe, "") && file.exists(conda_exe) ) {
+    return( conda_exe )
+  }
+
   bin_path <- file.path(install_root(), "miniconda", "condabin", c("conda", "conda.exe", "conda.bin", "conda.bat"), fsep = "/")
   bin_path <- bin_path[file.exists(bin_path)]
   if(length(bin_path)){
@@ -147,8 +166,23 @@ conda_bin <- function(){
 #' @rdname conda-env
 #' @export
 env_path <- function(){
+
+  current_env <- Sys.getenv("R_RPYMAT_CONDA_PREFIX", unset = "")
+  conda_exe <- Sys.getenv("R_RPYMAT_CONDA_EXE", unset = "")
+
+  re <- NULL
+  if(!identical(current_env, "")) {
+    re <- file.path(dirname(current_env), CONDAENV_NAME())
+  } else if( !identical(conda_exe, "") ) {
+    re <- file.path(dirname(dirname(conda_exe)), 'envs', CONDAENV_NAME())
+  }
+
+  if(length(re) != 1) {
+    re <- file.path(install_root(), "miniconda", 'envs', CONDAENV_NAME())
+  }
+
   return( normalizePath(
-    file.path(install_root(), "miniconda", 'envs', CONDAENV_NAME()),
+    re,
     winslash = "\\",
     mustWork = FALSE
   ) )
@@ -327,10 +361,10 @@ configure_conda <- function(python_ver = "auto",
     }
   }
 
-  if(dir.exists(path) && !force){
+  if(!conda_is_user_defined() && (dir.exists(path) && !force)){
     stop("conda path already exists. Please consider removing it by calling `rpymat::remove_conda()`")
   }
-  if(force || update || !dir.exists(path)){
+  if(conda_is_user_defined() && (force || update || !dir.exists(path))){
     miniconda_installer_url()
     tryCatch({
       reticulate::install_miniconda(path = path, update = update, force = force)
@@ -362,26 +396,52 @@ configure_conda <- function(python_ver = "auto",
   error <- FALSE
 }
 
+
+conda_is_user_defined <- function() {
+  actual_root <- normalizePath(file.path(conda_path(), ".."), mustWork = FALSE)
+  root <- normalizePath(install_root(), mustWork = FALSE)
+  !identical(actual_root, root)
+}
+
 #' @rdname conda-env
 #' @export
 remove_conda <- function(ask = TRUE){
-  root <- install_root()
   if(!interactive()){
     stop("Must run in interactive mode")
   }
-  if( !dir.exists(root) ){ return(invisible()) }
-  if( ask ){
-    message(sprintf("Removing conda at %s? \nThis operation only affects `rpymat` package and is safe.", root))
-    ans <- utils::askYesNo("", default = FALSE, prompts = c("yes", "no", "cancel - default is `no`"))
-    if(!isTRUE(ans)){
-      if(is.na(ans)){
-        message("Abort")
-      }
-      return(invisible())
-    }
-  }
 
-  unlink(root, recursive = TRUE, force = TRUE)
+  if(conda_is_user_defined()) {
+    envpath <- env_path()
+    if( !dir.exists(envpath) ){ return(invisible()) }
+    if( ask ){
+      message(sprintf("Removing conda at %s? \nThis operation only affects `rpymat` package and is safe.", envpath))
+      ans <- utils::askYesNo("", default = FALSE, prompts = c("yes", "no", "cancel - default is `no`"))
+      if(!isTRUE(ans)){
+        if(is.na(ans)){
+          message("Abort")
+        }
+        return(invisible())
+      }
+    }
+
+    system2(conda_bin(), args = c(sprintf("remove --name %s --all --yes", shQuote(CONDAENV_NAME()))))
+
+  } else {
+    root <- normalizePath(install_root(), mustWork = FALSE)
+
+    if( !dir.exists(root) ){ return(invisible()) }
+    if( ask ){
+      message(sprintf("Removing conda at %s? \nThis operation only affects `rpymat` package and is safe.", root))
+      ans <- utils::askYesNo("", default = FALSE, prompts = c("yes", "no", "cancel - default is `no`"))
+      if(!isTRUE(ans)){
+        if(is.na(ans)){
+          message("Abort")
+        }
+        return(invisible())
+      }
+    }
+    unlink(root, recursive = TRUE, force = TRUE)
+  }
 
   return(invisible())
 }
